1_RAG/requirement.txt
langchain 
langchain-google-genai
python-dotenv
langchain-core
chromadb 
langchain-community
pypdf
sentence-transformers
langchain-huggingface
langchain-chroma